fix/(CRED-289 CRED-284 CRED-277) Fix loc issues

+++ custom/crediblex/portfolio/loc/src/main/java/com/crediblex/fineract/portfolio/loc/data/LineOfCreditRequest.java
@@ -118,7 +118,7 @@ public class LineOfCreditRequest {
     private Integer cashMarginType;
 
     @Schema(example = "23", description = "Cash margin value")
-    private Integer cashMarginValue;
+    private Float cashMarginValue;
 
     @Schema(example = "1", description = "Interest charge time")
     private Integer interestChargeTime;
